working

+++ database/seeds/CategorySeeder.php
@@ -5,13 +5,21 @@ use Illuminate\Database\Eloquent\Model;
 class CategorySeeder extends Seeder{
 
 	public function run(){
-		
-		DB::table('categories')->delete();
+		if(DB::table('categories')->exists())
+			DB::table('categories')->delete();
 		$node = App\Model\Category::create([
 				'name' => 'Общие сведения о поставщике социальных услуг',
 				'slug'=>'obschie_svedeniya_o_postavschike_sotsialnyih_uslug',
 				'active'=>1,
 				'created_at'=>Carbon\Carbon::now(),
+				'children'=>[
+						[
+							'name' => 'Кадровое обеспечение',
+							'slug' => 'personal',
+							'active'=>1,
+							'created_at'=>Carbon\Carbon::now(),
+						],
+				]
 
 		]);
 		
@@ -20,6 +28,7 @@ class CategorySeeder extends Seeder{
 				'slug' => 'struktura_i_organy_upravlenija_organizacii_social_nogo_obsluzhivanija',
 				'active'=>1,
 				'created_at'=>Carbon\Carbon::now(),
+
 		]);
 		$node = App\Model\Category::create([
 				'name' => 'Формы и виды социальных услуг',
@@ -29,7 +38,7 @@ class CategorySeeder extends Seeder{
 		]);
 		$node = App\Model\Category::create([
 				'name' => 'Тарифы на социальные услуги',
-				'slug' => 'tarify_na_social_nye uslugi',
+				'slug' => 'tarify_na_social_nye_uslugi',
 				'active'=>1,
 				'created_at'=>Carbon\Carbon::now(),
 		]);
@@ -41,7 +50,7 @@ class CategorySeeder extends Seeder{
 		]);
 		$node = App\Model\Category::create([
 				'name' => 'Материально-техническое обеспечение социальных услуг',
-				'slug' => 'material_no-tehnicheskoe_obespechenie_social_nyh uslug',
+				'slug' => 'material_no-tehnicheskoe_obespechenie_social_nyh_uslug',
 				'active'=>1,
 				'created_at'=>Carbon\Carbon::now(),
 		]);
@@ -75,7 +84,7 @@ class CategorySeeder extends Seeder{
 		]);
 		$node = App\Model\Category::create([
 				'name' => 'Материально-техническое обеспечение социальных услуг',
-				'slug' => 'material_no-tehnicheskoe_obespechenie_social_nyh_uslug',
+				'slug' => 'material_no-tehnicheskoe_obespechenie_social_nyh_uslugi',
 				'active'=>1,
 				'created_at'=>Carbon\Carbon::now(),
 		]);
